统计修改

+++ mktcenter-serviceimpl/src/main/java/com/bizvane/mktcenterserviceimpl/service/impl/TaskServiceImpl.java
@@ -372,7 +372,7 @@ public class TaskServiceImpl implements TaskService {
                 }
                 ResponseData<CouponFindCouponCountResponseVO> couponCount= couponQueryService.findCouponCountBySendBusinessId(task.getTaskId(), sendType, sysBrandId);
                 CouponFindCouponCountResponseVO data = couponCount.getData();
-                Long couponSum = data.getCouponSum();
+                int couponSum = data.getCouponSum();
                 task.setOneTaskInvalidCountCoupon(Long.valueOf(couponSum));
 
                 allPoints=allPoints+task.getOneTaskPoints();
